+++ tests/algo_test.py
@@ -11,94 +11,6 @@ from neuronav.agents.mc_agents import QEC, QMC
 from neuronav.agents.dist_agents import DistQ
 
 
-def test_one_hot():
-    a = onehot(1, 5)
-    assert a.all() == np.array([0, 1, 0, 0, 0]).all()
-
-
-def test_two_hot():
-    a = twohot([1, 1], 3)
-    assert a.all() == np.array([0, 1, 0, 0, 1, 0]).all()
-
-
-def test_plot_value_policy():
-    env = GridEnv()
-    agent = TDQ(env.state_size, env.action_space.n)
-    plot_values_and_policy(agent, env, [9, 9], "Test Plot")
-
-
-def test_objects_graph():
-    objects = {"rewards": {0: 1}}
-    env = GraphEnv()
-    env.reset(objects=objects)
-    env.step(env.action_space.sample())
-
-
-def test_objects_grid():
-    objects = {"rewards": {(1, 1): 1}, "markers": {(1, 1): (1, 0, 0)}}
-    env = GridEnv()
-    env.reset(objects=objects)
-    env.step(env.action_space.sample())
-
-
-def test_graph_obs():
-    for obs_type in GraphObsType:
-        env = GraphEnv(obs_type=obs_type)
-        env.reset()
-        env.step(env.action_space.sample())
-
-
-def test_graph_structure():
-    for structure in GraphStructure:
-        env = GraphEnv(graph_structure=structure)
-        env.reset()
-
-
-def test_grid_orient():
-    for obs_type in GridObsType:
-        env = GridEnv(orientation_type=OrientationType.variable, obs_type=obs_type)
-        obs = env.reset()
-        env.step(env.action_space.sample())
-        if obs_type != GridObsType.index:
-            assert obs.shape == env.observation_space.shape
-
-
-def test_grid_obs():
-    for obs_type in GridObsType:
-        env = GridEnv(obs_type=obs_type)
-        obs = env.reset()
-        env.step(env.action_space.sample())
-        if obs_type != GridObsType.index:
-            assert obs.shape == env.observation_space.shape
-
-
-def test_grid_topo():
-    for topo in GridTopography:
-        for size in GridSize:
-            env = GridEnv(topography=topo, grid_size=size)
-            env.reset()
-
-
-def test_seed_gridenv():
-    env = GridEnv(seed=0, obs_type=GridObsType.symbolic)
-    env.reset(stochasticity=1.0)
-    obs_a, rew, don, _ = env.step(env.action_space.sample())
-    env = GridEnv(seed=0, obs_type=GridObsType.symbolic)
-    env.reset(stochasticity=1.0)
-    obs_b, rew, don, _ = env.step(env.action_space.sample())
-    assert obs_a.all() == obs_b.all()
-
-
-def test_seed_graphenv():
-    env = GraphEnv(seed=0, obs_type=GraphObsType.onehot)
-    env.reset(stochasticity=1.0)
-    obs_a, rew, don, _ = env.step(env.action_space.sample())
-    env = GraphEnv(seed=0, obs_type=GraphObsType.onehot)
-    env.reset(stochasticity=1.0)
-    obs_b, rew, don, _ = env.step(env.action_space.sample())
-    assert obs_a.all() == obs_b.all()
-
-
 def test_td_q():
     env = GridEnv(obs_type=GridObsType.index, grid_size=GridSize.micro)
     obs = env.reset()
@@ -261,3 +173,9 @@ def test_grid_episode():
     grid_env = GridEnv()
     grid_agent = TDQ(grid_env.state_size, grid_env.action_space.n)
     _, _, _ = run_episode(grid_env, grid_agent, 100)
+
+
+def test_plot_value_policy():
+    env = GridEnv()
+    agent = TDQ(env.state_size, env.action_space.n)
+    plot_values_and_policy(agent, env, [9, 9], "Test Plot")
